requirements.txt
yt-dlp>=2024.1.0
openai-whisper>=20231117
openai>=1.0.0
anthropic>=0.20.0
supabase>=2.0.0
python-dotenv>=1.0.0
requests>=2.31.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
folium>=0.15.0
tqdm>=4.65.0
torch>=2.0.0
